Skip persisting replica token if the disk is unavailable(I/O errors) (#1313)
Skip persisting replica token if the disk is unavailable(I/O errors)

The background replica token persistor should skip bad disk if there is
hardware issues with it. If there is a disk I/O error, store will be
automatically shutdown first. And then, by checking if all the stores are
down on that disk, the persistor can determine if disk is unavailable
and skip it accordingly.