Fix netty ByteBuf release bug in the ambry-frontend (#1310)
ResponseInfo release the ByteBuf when GetOperation finishes handleResponse,
and it would effectively put ByteBuf back to the pool. But GetOperation might
buffer those bytes in a BlobDataReadableStreamChannel so releasing the
ByteBuf would eventually corrupt the response if the same ByteBuf is checked
out to hold different bytes.

This PR fixes this bug, but retaining ResponseInfo and put it in a map and later
release it. The reason to add retain method is to make logic easier to understand.
After this PR, ResponseInfo would be released after GetOperation finishes
handleResponse, but if it's retained, it will then again get released after
BlobDataReadableStreamChannel finishes writing to HttpResponse.