Add a NettyByteBufLeakHelper to help test ByteBuf leaking (#1311)
ByteBuf resource leak detection is currently relying on reading the test log
and find out whether there is any leaking log. This is not very reliable and
very tedious to do.

This PR provides a helper class to detect the ByteBuf leak. This helper class
relies on metrics from Netty library. It first disables the caches (caches would
mess up the deallocations), then it checks after test, if the number of active
allocations remain the same. If not, it fails the test.