Support getting end position of last PUT in BlobStore (#1307)
The PR introduces a new method in PersistentIndex to get end position of
last PUT record in BlobStore. Note that, the "last" PUT record is valid
only when the method is invoked. That is, there could be new PUTs
afterwards. The intention here is to help replica to check if its peer
replicas have caught up with latest PUT record when PUT is disabled on
this replica (this happens when current replica is being decommissioned)