WIP: Allow PutOperation to use dynamic success target
In the intermediate state of "move replica", when decommision of old
replicas is initiated(not transit to INACTIVE yet), the PUT requests
should be rejected on old replicas. For frontends, they are seeing both
old and new replicas(assume 3 old and 3 new), the success target should
be 6 - 1 = 5. In the aformentioned scenario, PUT request failed on 3 old
replicas and we are supposed to fail whole PUT operation. However, PUT
request is highly likely to succeed on 3 new replicas and we actually
could consider it success without generating "slip put" (which makes PUT
latency worse)
The new success criterion is:
 success cnt > Math.max(parallelism - disbled cnt - 1, 2)
The 2 here is original PUT success target, which is the lower bound of
PUT operation target.