Fix potential empty routing table snapshot during HelixClusterManager initialization
After routing table provider is created, initial instance config change
event is added into the queue. However, when the event will be dequeued
is not able to be determined. There could be an edge case where even
after clustermap is initialized, the initial event is still in the
queue which causes routing table snapshot to be empty. This PR explictly
checks the snapshot and waits until it gets populated. Otherwise, there
may be no available replicas for router to send requests to.