WIP: Implement replica Inactive-To-Offline transition on server side
The major change in this PR is to implement Inactive-To-Offline logic
during the transition. Specfically, there are 3 steps to go:
1. set local store state to OFFLINE and initiate disconnection
2. wait until enough peer replicas have fully caught up with local
store (include DELETE/TTLUpdate records)
3. storage manager stop the store
4. (TODO) update instanceConfig in Helix (will be in a separate PR)
In addition, this PR also allows store to create a decommission file to
indicate if local store is in decommission process. This is to
distinguish such store from regular stores and it also helps failure
recovery.