Minor changes to support recovering from replica addition failure (#1382)
This PR adds minor changes to resume dynamic replica addition if it
failed last time before updating InstanceConfig(clustermap update in
Helix). The logic is simple: if diskManager finds there is a existing
dir associated with replica to add, it first deletes it and recreates a
new one. (The files in previous dir may not be reliable)