Remove getMessageInfo from MessageStoreHardDelete interface (#1391)
Remove unnecessary method getMessageInfo from MessageStoreHardDelete interface.

When the log files are not segmented, each delete index entry has a field "originalMessageOffset" pointing to the original Put record, so that when fetching content for a deleted blob, we don't have to search the Put record from the index again, we simple read the content from the Put record located by the "originalMessageOffset".

This is not true after we segmented log files and start compaction. Now a Delete index entry might not have a "originalMessageOffset" when the Put entry is not located in the same segment as the delete entry, so we have to search the Put entry from the index again not matter what. This totally obsolete the "originalMessageOffset" and thus the logic to read the content from the "origianlMessageOffset".