Add precondition to metadata updates to fail on conflicts (#1380)
When we update blob metadata in ABS and Cosmos, we have to first pull the whole metadata record (a property map in ABS and a document in Cosmos), update the field of interest, and upload the modified record. (There is no single-field update operation.) In live-serving mode, both the ambry-frontend and ambry-vcr may try to update the same blob metadata concurrently. Hence, it is possible for updates to get clobbered due to race conditions. To prevent this, we make use of the resource etag, along with the If-Match request condition, to fail conflicting updates and force them to retry.

Note: the mechanism used to test conflicts is non-standard as it relies on a callback in the production code for the test to inject a conflicting modification. I'm open to suggestions for cleaner alternatives.