Change Store interface to not use MessageWriteSet (#1386)
delete and updateTtl method in interface Store take MessageWriteSet as its only parameter. This is not ideal since the data from MessageWriteSet would be persisted in the LogSegment files on disk, which means there is a chance that outside of the BlobStore implementation, we can construct any arbitrary bytes and persist them in the log files. This is not very bad, for we control how to construct those bytes, but from interface's perspective, it's not a good one.

And in order to support persisting undelete record, we need to get the lifeVersion from the PersistentIndex before we can construct a UndeleteFormatInputStream. The same goes to delete and ttlUpdate record. But in AmbryRequests, we don't yet know the lifeVersion before creating FormatInputStream for those record because of Store interface requires MessageWriteSet as the parameter. Changing the interface to take MessageInfo not MessageWriteSet solves this problem.

This PR changes that to make sure BlobStore would construct all the delete and updateTtl records by itself. We should change put method as well, but later.