Make PartitionSelectionHelper incorporate clustermap changes (#1405)
Previously, data structures in PartitionSelectionHelper are constant
after initialization. With "Move Replica" feature on, this should
dynamically incorporate any changes in clustermap (i.e. new
replica/partition added or replica is removed). In this PR, we make it as
a listener of clustermap and re-populate all in-mem data structures if
number of replicas/partitions has changed.