Add new version to MessageInfoAndMetadataListSerDe to delivery lifeVersion on wire (#1424)
currently we don't serialize lifeVersion in message info at GetResponse and ReplicaMetadataRespone.
This PR adds a new version to MessageInfoAndMetadataListSerDe class to pass lifeVersion on wire.

After this is merged, we have to first deploy it to all the frontend so the frontend can deserialize newer version. Then we have to deploy to all server, so server can deserialize newer version.

Then we can push a new PR to write message info with newer version from ambry server.