Fix the null pointer exception in hard deleter (#1426)
This should fix the null pointer exception in hard deleter that recorded in this issue #1425

The NPE is caused by performing recovery. When we persist HardDeletePersistItem to the disk, we only write the BlobReadOptions and the recoveryInfo bytes, without the start token. So when we perform the recovery, we read the BlobReadOptions and the recovery info bytes out, but now the HardDeletePersistItem doesn't have a start token. Luckily, we don't need start token for recovery, we only have to use the BlobReadOption and recovery info bytes. After performing the recovery, we have to clear the hardDeleteRecoveryRange, so later all the items added to the list would have all three fields.

Before the PR, the hardDeleteRecoveryRange is not cleared after performing the recovery, so next time when calling pruneTill method, some of the item doesn't have a startToken. Clearing it would fix this issue.