Revert the message header version back to 2 and fix a hard deleter race condition (#1430)
Bump the version from 2 to 3 without deploying a version of ambry server that knows how to transform version 3 message is a mistake. And it causes replication to fail.

When we deploy a new version of ambry server that has version 3 to one machine, the other machines doesn't know how to transform the message in version 3. So in replication, they all fail. The right way of changing version should be

Add version 3 to all transformers, open source and close source.
Deploy this version to all ambry servers
Change the version to 3
Deploy the new version to all ambry servers.