Remove dots from java package directory names
One odd issue I have always seen is that when depending on ambry as a
library, intellij is unable to associate the source files with package
files even though both jars are present. I found that the reason for
this is because we include dots (.) in our package directory names
instead of the traditional way of having a separate directory for each
package name part.

This commit replaces the directory names using a script
(e.g. com.github.ambry becomes com/github/ambry)

The script used for this (remove-dot-dirs.sh) is included in this commit
for others to use to fix their open PRs after merging