Add support for cloud replicas to OperationTracker
CLOUD_BACKED replicas differ from DISK_BACKED replicas in that they
only require one succesful response to deem the operation a success.
The changes in this commit tune the success target and parallelism based
on the type of replica requests are sent to. For DISK_BACKED replicas,
the policies remain the same, but for CLOUD_BACKED replicas, one
successful response will complete the operation. Additionally, only one
request will be sent by SimpleOperationTracker at a time when a cloud
request is in flight (since this one request would be able to
successfully complete the operation).