Fix NPE in router metrics when handling newly added nodes (#1446)
There is a NPE issue in GetBlobOpertion/DeleteOperation/TtlUpdateOperation
due to node level metic not found in router metrics. This happens when we
dynamically add new nodes/partitions to cluster. The frontend is able to
get new replicas from new nodes but can't find NodeLevelMetric
associated with the new nodes that host these new replicas and
NullPointerException occurs afterwards. To fix this, we allow router
metrics to dynamically add new node metric if it's not present
previously.