Fix ReplicaSyncUpManager updates lag of replica in wrong state (#1464)
There is a bug in replica thread where we update replication lag for
bootstrap replica in ReplicaSyncUpManager. It relies on whether there is
a localReplicaLagInfo present in ReplicaSyncUpManager to update
replication lag of certain bootstrap replica. However, if a replica is
being decommissioned, it also creates a localReplicaLagInfo in
ReplicaSyncUpManager. The replica thread may mistakenly updates the
replication lag of a decommission replica.
This PR adds additional check to ensure the current state is BOOTSTRAP
before updating replication lag in ReplicaSyncUpManager.