Increase retention time to fix the hard deleter test failure (#1468)
This PR increases the retention day in the hard delete integration test to fix test failure.
The problem is when we delete a blob in the test case, we undelete it later. There are some time gap between these two operations and somehow the time instance runs too far to surpass one day, which is the retention day for deleted blob. So when we undelete it, it's already too late.
This PR changes the retention day to 7 and hopefully this would fix this the failure.