Treat 'SSLEngine closed already' as a client termination exception. (#1466)
When clients close the connection from their side, the router may
still be delivering data. When using SSL, the write call in
NettyResponseChannel may encounter an exception indicating the SSLEngine
has already been closed, which can happen when the connection is in the
process of being closed.

This change treats such exceptions as client-side connection
terminations so they will not be logged so verbosely. See
NettyResponseChannel::log for how Utils::isPossibleClientTermination
is used.