Add basic support for multiple participants on server side (#1472)
In some cases where single datanode has to parcipate into mulitiple clusters
and it needs to update node info in these clusters separately. This requires
muptiple participants on same node and each of them is responsible for interacting
with corresponding cluster. With this feature, we are able to build a mirror cluster
that helps us perform some monitoring and adminstration operations (i.e. ZK migration).