Use conservative findEntriesSince size limit when update and put are both in the journal (#1484)
In the case where both a put and a TTL update show up when finding
entries in the journal, the most recent index entry found will always be
the ttl update entry. This means that the size of the TTL update entry
will be checked against the size limit of the findEntriesSince query.
However, a replicator will often have to fetch the original blob
anyways if they don't already have the put entry, which means that
they could do a much larger batch get than expected, leading to
excessive memory consumption on the batch-get server side.

To addresss this issue, this commit adds logic to check if there is a
put that the caller of findEntriesSince has likely not encountered yet,
and, if so, counts its size towards the limit. This behavior is not used
for deletes, since there is no need to fetch a deleted blob's data.