Support generating partition override map from config file in bootstrap tool
Previously, we have to manually edit state of partition in static clustermap. If
there are thousands of partitions to be marked as ReadOnlly, it will take some
effort to change clustermap(either by script or manually). This PR allows user to
feed tool with a static config file that contains a list of partition ids to
override. These partitions are separated by comma and bootstrap tool will check
if id is valid (i.e. numeric, within valid range).