Fix a MessageSievingInputStream test failure (#1491)
The root cause for this failure is we create a corrupted data stream for message info 2, and sometimes the header version is valid since it's all a random number. When the header version is valid, it will might fail because of StoreKey is not valid. Since when the header version is valid, the entire header bytes are valid, and store key is the item following the header, and it will fail.

In this PR, we start verifying the header's sanity after reading the header out. This will make header invalid even if the header version is valid.