Introduce helixAccountService to Ambry Server and get Containers by status (#1482)
This PR supports below part of stages for container deletion compaction phase.

Implement getContainersByStatus in helixAccountService, so it can query DELETE_IN_PROGRESS containers from ZooKeeper.
Introduce helixAccountService to Ambry Server, so BlobStoreCompactor is able to get containers by status through helixAccountService.
Add unit test to check getContainersByStatus() in helixAccountServiceTest.
Please reference the detail below:
Compaction is the place to delete blobs from DELETE_IN_PROGRESS containers.
Compaction runs in cycles and it copies valid data to new log segments in each cycle. The valid data is defined as [“not deleted” && “not expired” && “still in the retention time window if deleted or expired”]. We expand the definition of valid data to [“not deleted” && “not expired” && “still in the retention time window if deleted or expired” && “not from DELETE_IN_PROGRESS container”]. Then all blobs in DELETE_IN_PROGRESS are considered as invalid data and should be compacted.