Allow router to send cross-colo request to dc with most replicas
In a case where a certain dc is decommissioned and blobs previously uploaded
to this dc now have a unrecognizable dc id. Clustermap will treat originating dc
of this blob as null. When request didn't succeed in local dc, operation tracker
should be allowed to try on remote dc with most replicas first so as to improve
success rate of cross-colo requests(i.e GET/DELETE etc). This is useful in cluster
with "unbalanced" replica distribution (i.e. 3 replicas in local dc and 1 replica
per remote dc).