Make verifySignedRequest method async (#1509)
UrlSigningService::verifySignedRequest() may need to call remote APIs to
verify a signature or token. If the remote service is degraded and the
calls are slow, executing them synchronously can block handling of new
incoming requests (since it is often called from preProcessRequest,
which runs on the AsyncRequestWorker thread).

This PR makes the verifySignedRequest method callback/future based to
allow for implementations to make such requests asynchronously.