Make partition class more permissive for migrations (#1486)
If a user is migrating between ambry clusters where the partition classes
are different, they may run into a situation where they either change
replicationPolicy on their container first, leading to puts in the
original cluster fail, or change the cluster used first, which leads to
puts in the new cluster failing. To aid in migration, this PR changes
the partition class selection behavior to fall back to the default
partition class. In this case, a metric will be emitted and a warn log
message will be written.

Additionally, this PR treats an empty string in replicationPolicy the
same as a null string, as this case was ambiguous before.