Adding config to specify the type of replication model to be used for inter colo replication (#1512)
Adding config to specify the type of replication to be used for inter colo replication. It has two values - "all-to-all"(current model) or "leader-based" with default value is "all-to-all". If set to "leader-based", inter colo replication will be limited to the leaders(as elected by Helix) of the partition of each datacenter. If set to "all-to-all", inter colo replication will be in an all-to-all fashion, i.e. each replica talks to all other replicas irrespective of their state. Intra colo replication will continue as all-to-all fashion in both the models.