Fix a already deleted error (#1520)
When frontend sends a delete request to all servers, some of them would respond faster than others. If at the same time replication thread just happens to replication the delete request from the already responded server to the local server, there might be a conflict.

Since replication thread would also delete a blob with a valid lifeVersion, we should compare the state of the blob when lifeVersion of this blob equals to the lifeVersion from message info. And when they are the same, we should throw an ID_DELETED error instead of LIFE_VERSION_CONFLICT.