Ignore ttl update operation time in replication (#1533)
This PR ignores the operation time checking in ttl update when the request is coming from replication. The reason for not checking the ttl update's operation time, is that the operation time in the message info might not be the real operation time issued from frontend anymore.

For instance, frontend can send PUT, TTL_UPDATE to a server1 and only PUT to server2. And somehow server2 goes offline for a long time. Within this period, frontend sends DELETE to server. Assume this delete's operation time is after original PUT's expiration time, then server2 would have trouble to replicate TTL_UPDATE and DELETE from server2.

When server2 goes online again, it gets MessagInfo from server1 about this blob. The MessageInfo would have isDeleted=true, isTtlUpdated=true, operationTime=deleteOperationTime. This is obviously correct, since the blob is already deleted and ttl updated in server1 and the delete is the latest operation. At this moment, server2 would try to apply ttl update to local blobstore, with operationTime = deleteOperationTime. We know this would fail, since deleteOperationTime is after the local PUT's expiration time.

This PR solves this particular issue by ignoring the operation time when it's replication.