Fix a concurrency bug caused by offset comparison (#1548)
We are seeing in EI, some partition has duplicated delete record in log, but not in index. After investigation, this is caused by an offset comparison, which is fixed in this PR.

Brief explanation:

There are two concurrent delete requests, one of which from frontend, the other from replication.
Both delete requests will call delete method, and both would check the validity of delete record up to line 553, and both would be valid. (There is no delete in index yet)
After that, one of delete would enter the mutual exclusive block starting from link 553, and persist delete record to log and index
The second delete would figure the indexEndOffsetBeforeCheck is not the same as currentIndexEndOffset, assuming a delete record is 100 bytes, and indexEndOffsetBeforeCheck is 1000. then after first delete record, currentIndexEndOffset should be 1100
Second delete call gets the the first delete record, and it's offset should be 1000. so here the comparison should be >=, rather than >.