Optimizations for BlobStoreStats
There seem to be a couple hot spots for stats collection. The main one
is querying the full index for each put index entry in
updateExpiryTimeForAllPuts. I do not have a way of reducing the number
of index lookups here yet, but I am trying to find a couple places where
we can improve things along the way:

In this commit, I get rid of one intermediate list of valid index
entries and instead directly update the map instead of iterating an
extra time and allocating an extra ~40k item list.