Make storage manager register state change listener in multiple participants
This is a minor improvement for ZK migration. Previously, only primary participant
registers storage manager listener and other participant has no listener. The side
effect is, if store is not started due to disk I/O error(during startup process),
the second participant is not able to mark this store in ERROR state as it has no
listener and is no-op in OFFLINE->BOOTSTRAP transition. This PR ensures both participants
are able to detect store failure during startup.