Patches for Http2 (#1551)
Provided an initializer when stream channel is created. This allows setup pipeline first and then return stream to claimer. Without this, it may face a timing issue that claimer try to take action on pipeline but pipeline is empty.
Add metrics and logs for request to drop.
Polish releaseAndCloseStreamChannel.
Set REQUEST_INFO to null whenever possible to reduce reference to object.