Optimizations for BlobStoreStats (#1549)
There seem to be a couple hot spots for stats collection. The main one
is querying the full index for each put index entry in
updateExpiryTimeForAllPuts.
This commit gets rid of one intermediate list of valid index
entries and instead directly updates the map instead of iterating an
extra time and allocating an extra ~40k item list.