Make loggers static final when possible
In almost all cases, loggers should be static. When run with log4j2,
there is more overhead to get a Logger instance, whereas with log4j1,
usually on the first request to get a logger for a given name would be
expensive. This means that instantiating loggers for classes with many
instances such as BlobReadOptions, which is instantiated for every blob
served, have a visible impact when analyzed with a sampling profiler.

This commit is a mass find and replace of non-static loggers.