Change sanity check logic for undelete so recovery can work (#1570)
There are several changes to this PR

Undelete will no longer check if the blob is expired for replication
Undelete will now take a MessageInfo as argument in recovery, and will only use info to construct an index value, without any sanity checking.
copyRecords in compaction will not use markAsUndelete to check sanity, instead it will have it's own sanity check.
This would fix certain recovery and replication corner cases.
Case 1. We have P, D, U, T for a blob. The PUT has a short ttl, so we need the last TTL_UPDATE to make it permanent. When the whole index files are gone and we are reconstructing index from log segments, we would fail on U since the P is expired in recovery. This PR would fix that.

Case 2. We have P, D, U, D for a blob. The compaction compacts Put and first DELETE, (UNDELETE and second DELETE are in a different segment, which is not under compaction). Then we only have U and D after compaction. When the whole index files are gone and we are reconstructing index from log, we would fail on U since there is nothing prior to it. This PR would fix that.

Case 3. We have P, D in local, and the remote has P, D, U, T. Somehow local stops for a while until P is expired. When restarting local process, it starts replicating from remote. And it will fail on applying U, since P is expired. This PR would fix that.