Introduce PropertyStoreToDataNodeConfigAdapter
Introduce a DataNodeConfigSource implementation that can be used to
listen for changes to configs using HelixPropertyStore.

To avoid blocking the threads used by helix to send property store
change events, this implementation uses a background thread to read
records from the property store and notify any downstream listeners.

To prepare for future participant changes this also adds set and get
methods to DataNodeConfigSource for reading and updating
DataNodeConfigs.