Use DataNodeConfigSource in participant
Use DataNodeConfigSource as an abstraction around HelixAdmin and
InstanceConfigs in HelixParticipant to allow for future configurable
implementations to be plugged into this class.

For now, this only supports InstanceConfigToDataNodeConfigAdapter, but
will be modified to allow configuration in future commits.

This commit switches HelixParticipant to use a shared spectator
HelixManager instance for any config manipulation. This
avoids the need to construct a temporary HelixAdmin before the
"official" admin is ready.