Add one integration test for undelete index recovery (#1578)
Add one more integration test for undelete, it's testing the index recovery of undelete records.

It sends some records, including UNDELETE, to one of the servers and stops the server, then 
remove the index files and restart the server process. The restarted server process should recovery 
the index files from the log files.