Fix NPE issue when starting a stopped replica after reboot
There is an edge case where we successfully stop a replica and shut down the node
for some reason (i.e. deployment). After node is restarted, the stopped replica is
not able to start as storage manager checks its existence on stopped list. Replication
manager therefore doesn't create RemoteReplicaInfo for this replica. This brings an
issue that when we restart store because remote replica will send metadata request and
there is not corresponding RemoteReplicaInfo in ReplicationManager, which causes NullPointerException.
This PR ensures replica is added to ReplicationManager (if not present) when starting
the store. Also, it resets the partition state to OFFLINE to bring replica out of ERROR
state (note that the replica was put into ERROR state during server startup).