Change java ByteBuffer to netty ByteBuf in request and response. (#1579)
Starting use netty ByteBuf to replace java.nio.ByteBuffer in the network layer.
There are some issues with GetResponse, since it supports non-prefetched data from disk. Something need to be done in order to actually support ByteBufHolder for GetResponse.

After this PR, Request and Response would have a method "content", if the returned value from method content is not null, then caller can use the returned ByteBuf as serialized bytes for the request and response.