Drop http2 request in two cases (#1589)
Drop request on router requests to drop a request.

Drop request on writeAndFlush exceeds threshold.

Adjusted Netty High and Low WaterMark for isWriteable() tracking.

There was an assumption that a stream channel can be released twice. 
This is wrong because MultiplexedChannelRecord maintains a counter for 
streams on it. This PR fixs it by checking if RESPONSE_INFO attr null or not.