Fix reference counter error in socketnetworkclient (#1601)
SocketNetworkClient would throw lots of IllegalReferenceCountException when any ambry-server dies. 
This is caused by a bug where I released the request twice. When a request is sent to server completed, 
it will be released by selector. And when it gets disconnected when it's waiting for response, it will be 
released again. This PR fixes that. It checks if the request is sentCompleted, if not, it will release the 
request when disconnected.