FindKey method takes FileSpan as the entire search space. (#1558)
Before this PR, findKey in PersistentIndex, uses FileSpan in a unorthodoxy way. It basically uses FileSpan as a way to locate index segments to search.

For instance, if we have three index segments, and their respective start offset and end offset are [0, 1000), [1000, 2000), [2000, 3000), then when we have a FileSpan whose start and end offset are [500, 2500], findKey would know that the all three index segments should be searched up, since the start offset is located within the first index segment and the end offset is located at the third segment.

The problem the actual record's offset is 2700, which is not covered in the FileSpan, but covered in the third segment, this record will be returned by findKey. In this case, FileSpan doesn't serve as a search space for findKey, but rather an pointer to find the index segments, which then would become search space.

This PR would fix this issue by changing findKey and strictly respect FileSpan as search space. It also has a few subtle changes.

FileSpan's start offset is inclusive, but end offset would be exclusive.
Compaction without undelete, would persist a cutoff offset to indicate if the delete is in effect. The cutoffset used to be the start offset of the index segment, now it will be the end offset.