[SEGMENTED_GET_FIX] Return correct size in headers for segmented blobs. (#1616)
The size of the segmented blob can only be determined when the segmented chunk has been fetched. This PR fixes the issue of returning correct for segmented blobs by taking advantage of the fact that for a segmented blob, first chunk will always be a metadata chunk, and subsequently there will be only data chunk that will ever be returned.