[DataNodeConfigSource] use dedicated HPS instance
The property store provided by the HelixManager.getPropertyStore()
method does not support caching/subscribed paths. As a solution, we can
move the ownership/instantiation of the PropertyStore into the
PropertyStoreToDataNodeConfigAdapter class.

To keep views of the property store consistent between HelixParticipant
and HelixClusterManager, this commit also adds a shared pool of
DataNodeConfigSources so that the same instance (and ZkClient) can be
used by both classes.