[named_blobs] Add NamedBlobDB interface/MySQL impl
Add an interface and a MySQL implementation of NamedBlobDB.
This interface will enable interaction with a mutable mapping layer for
named blobs.

Add a mysql based implementation that supports the put/get/list/delete
APIs. For the first iteration, this does not include retry and failover to
other active databases. This implementation uses Hikari for connection
pooling and built-in mysql connector options for prepared statement
caching.