[BlobStoreCompaction] Always enable target index duplicate checking to ignore duplicate PUTs (#1660)
We experienced a compaction error due to duplicate PUTs in log. This is a weird error since there shouldn't be any duplicate PUTs in the log. However, it takes too long to figure out why it happened. Before we fully understand the root cause, this is a bandage to temporarily fix the issue.

We can always check the duplicates in target persistent index while we are check if a IndexValue is a valid one or not. If it's a duplicate of certain IndexValue in target persistent index, then it's not a valid one. This logic would ignore the duplicate PUT and mark it as invalid.