Fix null replica in HelixClusterMap when decommissing replicas.
Ambry frontend now allows GET operation to try on OFFLINE replicas. During replica decommission, the first step is to
disable replica. This will eventually bring replica to OFFLINE state and remove its entry from DataNodeConfig. All listeners
(both frontends and servers) will receive the DataNodeConfig change and remove this replica from in-mem clustermap. However,
at this point of time, the IdealState in Helix hasn't been changed and this replica still exists in external view. So when
router calls getReplicaIdByState(), the RoutingTableProvider still returns instance(server) associated with this replica. This
causes a null replica when router queries in-mem clustermap by this instance and our code doesn't invalidate the null value in
the return result, which results in NPEs.
This PR makes two changes:
1. Add a config to control whether to add down replicas to the replica pool in operation tracker. If yes, down(offline) replicas
   will be added to the end.
2. Preclude null replicas in the method of getReplicaIdsByState().