[named_blobs] Avoid bad affected row count assumptions
Previously, the put code used an "insert on duplicate key" expression
with conditional assignments. However, this required using the affected
rows to count how many rows actually changed. If the DB is set up to use
triggers that assign other metadata columns, the row may change even if
the specific fields set in the update call did not.

As an alternative, we can switch to the default "found rows" jdbc
behavior and use an optimistic insert followed by a conditional update.