Add LogSegmentName class (#1752)
Parsing the log segment name to find the position and generation becomes
expensive when used as a comparator in frequently used index data
structures. This commit adds a class that represents a parsed log
segment name with a lighter weight compareTo method.

It is a significant refactoring to replace all usages of log segment
name strings with this class, so I only adopted the new class in Offset,
which should help in the frequently used
PersistentIndex.validIndexSegments and Journal data structures.