Integrate request quota with frontend. (#1763)
This PR applies the request quota to frontend request flow. Note that in order to complete the integration with frontend in open source code two more components are required which will come as future PRs in the interest of keeping this PR short. The two components are

Returning quota response headers containing quota usage and request costs, to response sent from frontend to client.
Integrating storage quota with the quota interfaces.