Hide Stats Key prefix in Utils class (#1777)
In stats reports, we use string as the key for partition id, account id and container id, with some prefix. 
Eg. for account id 100, the account key is A[100]. The prefix is spread all over the codebase. 
This PR reuse the common methods in Utils class to replace those statements that directly implement 
the logic of converting ids to string keys and vice versa.